Issue #2472767: Convert tags into plugins
This is the first pass of a huge overhaul of XBBCode,
moving from D7 hook-based discovery to annotated plugins.

Large parts of the code are directly based on the core filter API,
being that an XBBCode filter instance is a collection of tag plugins
in the same way a text format is a collection of filter plugins.

This is very unfinished. Working components include:
- Plugin discovery and configuration (both global and format-specific)
- Filter processing and filter tips.

Missing components include:
- Custom tags
- Twig integration [#2474505]