Issue #2473231: Run the whole module through drupal/coder.
This patch fixes a host of whitespace and code-style errors,
implements variable naming conventions, and fully completes
and corrects the PHPDoc comments.

Note that the annotation comment in TableTagPlugin cannot validate:
The Drupal style guide requires each line to begin with "*",
but the Symfony annotation parser would include these in the string.
The linebreaks themselves are required by the BBCode tag format.