Issue #2629674: A view custom tag operation.
The view operation displays a form with disabled or read-only
fields, including the contents of an external template file.
It is linked on the tag overview for all tags that are not editable.

This patch also moves the "edit" operation down to "/edit",
and makes the text areas load with the correct size to fully
show their contents.