RegEx fix: Escape quotes in unquoted values.
Unquoted values must escape *all* quotes, to ensure that
they are not mistaken for quoted values if the same quotes are
at the beginning and end of the value.

This patch also simplifies the regular expression, to replace
/([^\\\\]|(\\\\\\\\)*\\\\)*"/ with /([^\\\\"]|\\\\[\\\\"])*/

Bugfix in Element, where escaped spaces in unquoted values were
not actually unescaped after parsing.