Issue #2634550: More robust processing.
If other filters are applied before BBCode rendering,
then the encoded source may not be at the end of the string anymore.
It must be found and cut out of the text.

This patch also ensures that other [-*xbbcode][-*/xbbcode] tags in
the text are masked by an additional "-" before processing.