Code cleanup: TagPluginCollection::has().
Even though this override does nothing, it is needed in order to let
its parent method (with no return type) fulfill the contract of its
interface (which returns bool).

To throw off the PHP CodeSniffer (which does not understand this, see
https://github.com/squizlabs/PHP_CodeSniffer/issues/2230), add a (bool)
cast. This also does nothing, but it looks cute and clears validation.