Allow the HTML filter to be disabled.
This step is a safety measure to ensure that even in the absence of
other filters, the module will not emit unsafe markup.

However, in some use cases a prior filter may deliberately add
markup that would be unsafe if entered by users.

An extra option now allows this filter to be skipped, while explicitly
warning about the risk of doing so.