Add "Copy" button on tag forms.
Pressing the button-style link opens the copy form for the current tag,
which is mostly useful when viewing a non-editable tag.
On edit forms the button is labeled to warn about unsaved changes.

Slightly clean up the logic on the read-only form, by adding an #access
check to the submit button instead of disabling it in the form builder.