Expose tag attachments in edit form.
Internally, tag entities have always had support for page attachments,
but these were not exposed in the form, and could only be set by a module
or by editing the site configuration directly.

This also expands test coverage to check for the inclusion of attachments
on the tag edit form.